Split addon helper and loader
The addon helper is responsible for providing convenient functionality
to addons while the loader instantiates and manages the addons.
