+++ manimlib/__init__.py
@@ -4,7 +4,7 @@ import manimlib.config
 import manimlib.constants
 import manimlib.extract_scene
 import manimlib.stream_starter
-import manimlib.addon_helper
+import manimlib.addon_loader
 
 
 def main():
@@ -15,13 +15,13 @@ def main():
     else:
         print("Starting...", end="\r")
 
-        manimlib.addon_helper.read_addons(True)
+        manimlib.addon_loader.read_addons(True)
         args = manimlib.config.parse_cli()
-        
+
         if not args.livestream:
             config = manimlib.config.get_configuration(args)
             manimlib.constants.initialize_directories(config)
-            manimlib.addon_helper.pass_config_to_addons(config)
+            manimlib.addon_loader.pass_config_to_addons(config)
             manimlib.extract_scene.main(config)
         else:
             manimlib.stream_starter.start_livestream(
