+++ manimlib/addon_info.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python
-import manimlib.addon_helper
+import manimlib.addon_loader
 from threading import Timer
 import sys, time
-
 
 
 sys.stdout.write( 'Loading addons...\r' )
 sys.stdout.flush()
 
-manimlib.addon_helper.read_addons(True)
-manimlib.addon_helper.print_addon_info()
+manimlib.addon_loader.read_addons(True)
+manimlib.addon_loader.print_addon_info()
