+++ manimlib/extract_scene.py
@@ -127,7 +127,7 @@ def get_scene_classes_from_module(module):
     ]
 
 
-def main(config):
+def main(config={}):
     module = config["module"]
     all_scene_classes = get_scene_classes_from_module(module)
     scene_classes_to_render = get_scenes_to_render(all_scene_classes, config)
@@ -147,13 +147,13 @@ def main(config):
     for SceneClass in scene_classes_to_render:
         try:
             # Before we start rendering, fire the on_render_ready() function in each addon
-            manimlib.addon_helper.run_on_render_ready(scene_classes_to_render)
+            manimlib.addon_loader.run_on_render_ready(scene_classes_to_render)
 
             # By invoking, this renders the full scene
             scene = SceneClass(**scene_kwargs)
 
             # By this point, the scene is done rendering, so fire the on_render() function in each addon
-            manimlib.addon_helper.run_on_rendered(scene_classes_to_render)
+            manimlib.addon_loader.run_on_rendered(scene_classes_to_render)
 
             open_file_if_needed(scene.file_writer, **config)
             if config["sound"]:
